fix(attention): wire up the SM120 NVFP4 LSE epilogue and fp32 qk_correction matmul
LSEWriter had no call site, so fwd returned an uninitialized lse buffer
(the existing tests only assert non-NaN, which held or failed depending
on allocator reuse). Call it from the consumer loop after the softmax
finalize, using the per-warp-group PV mma for the row mapping plus an
explicit row offset, and remove the fp8_scalexfp4_scale_log2 factor that
row_sum carries for the FP4 P quantization so lse is the plain
ln-sum-exp of the scaled scores. The dead run() forwarding wrapper and
its stale m16n8 fragment asserts are updated/removed.

Also compute the qk_correction matmul in fp32 (review feedback): a
float16 matmul output would overflow at 65504 even though the
accumulation runs in fp32.

Validated on RTX PRO 6000: uniform scores give lse = ln(seq_len) exactly;
iid/structured lse mean abs error vs exact logsumexp is ~0.01-0.02 (max
0.30 on causal edge rows). New test covers causal and non-causal lse.
AI-assisted.

+++ flashinfer/nvfp4_attention_sm120.py
@@ -125,7 +125,11 @@ def _preprocess_qkv(
     # seq_len / 128, seq_len]), or a single row when per_block_mean=False.
     # The kernel's TMA descriptor addresses the tensor this way and
     # broadcasts each row across the 128 rows of the Q tile in smem.
-    qk_correction = torch.matmul(qm, k.transpose(-2, -1)).to(torch.float32)
+    # Multiply in fp32: a float16 matmul output would overflow at 65504
+    # even though the accumulation itself runs in fp32.
+    qk_correction = torch.matmul(
+        qm.to(torch.float32), k.transpose(-2, -1).to(torch.float32)
+    )
     qk_correction = qk_correction.contiguous()
     return q.contiguous(), k.contiguous(), v.contiguous(), qk_correction
